Harden notebook content parsing and validation
